Add show page styling and reorganize show page
